checkmap
working on checkmap stuff and it's being annoying and copying files and not wanting to include things so gimme a few im working on it

+++ game0/DungeonHero.h
@@ -17,7 +17,7 @@ private:
 	bool hasFireball;
 
 	void kbd(const df::EventKeyboard* p_keyboard_event);
-	void move(int dir);
+	void move(float dir);
 	void advance(int dir);
 	void fire(df::Vector target);
 	void mouse(const df::EventMouse* p_mouse_event);
